frozen-dependency-ctf/requirements.txt
# pre-empt other packages trying to install a higher version of pyproj
# this one works, and higher versions cause ceaseless issues with docker
pyproj>=2
wtforms<=2.1

# Python framework
flask

# Accounts framework
flask-security-too==3.3.0
flask-sqlalchemy
bcrypt

# client used for the WMTS API
# this client is a forked version of the geopython OWSLib client
-e git://github.com/ectoglasses/OWSLib.git#egg=OWSLib

# Library for sqlalchemy database (un-needed for sqlite and datastore)
# sqlalchemy

# Library for GCP Datastore inegration (un-needed for sqlite and sqlalchemy)
# google-cloud-datastore

# Python WSGI HTTP server (currently un-used)
# gunicorn

